fix: #67 구인글 생성 API 수정
