+++ src/types/index.ts
@@ -1,16 +1,8 @@
 import type { SignUpPayload } from './payload';
 import type { UserReponse, User } from './response';
 import type { CustomInstance } from './api';
-import type {
-	TitleInfo,
-	OptionList,
-	Option,
-	Role,
-	InputRoleForm,
-	InputState,
-	InputCourse,
-	TempRole,
-} from './information';
+import type { TitleInfo, Option, OptionList } from './recruit/response';
+import type { InputState, RoleForPost, InputCourse, InputRoleForm, Role } from './recruit/payload';
 import type { Comment, CommentInputFunctions } from './comment';
 import type { JsxElementComponentProps } from './indexSigniture';
 
@@ -29,5 +21,5 @@ export type {
 	InputRoleForm,
 	InputState,
 	InputCourse,
-	TempRole,
+	RoleForPost,
 };
